node/approval-voting: Introduce framework for full subsystem tests
