+++ node/core/approval-voting/Cargo.toml
@@ -42,3 +42,4 @@ maplit = "1.0.2"
 polkadot-node-subsystem-test-helpers = { path = "../../subsystem-test-helpers" }
 assert_matches = "1.4.0"
 kvdb-memorydb = "0.10.0"
+rand = "0.8"
