+++ node/core/approval-voting/src/import.rs
@@ -557,7 +557,7 @@ pub(crate) async fn handle_new_head<'a>(
 }
 
 #[cfg(test)]
-mod tests {
+pub(crate) mod tests {
 	use super::*;
 	use crate::approval_db::v1::DbBackend;
 	use polkadot_node_subsystem_test_helpers::make_subsystem_context;
@@ -565,11 +565,11 @@ mod tests {
 	use polkadot_primitives::v1::{SessionInfo, ValidatorIndex};
 	use polkadot_node_subsystem::messages::AllMessages;
 	use sp_core::testing::TaskExecutor;
-	use sp_runtime::{Digest, DigestItem};
-	use sp_consensus_babe::{
+	pub(crate) use sp_runtime::{Digest, DigestItem};
+	pub(crate) use sp_consensus_babe::{
 		Epoch as BabeEpoch, BabeEpochConfiguration, AllowedSlots,
 	};
-	use sp_consensus_babe::digests::{CompatibleDigestItem, PreDigest, SecondaryVRFPreDigest};
+	pub(crate) use sp_consensus_babe::digests::{CompatibleDigestItem, PreDigest, SecondaryVRFPreDigest};
 	use sp_keyring::sr25519::Keyring as Sr25519Keyring;
 	use assert_matches::assert_matches;
 	use merlin::Transcript;
@@ -650,7 +650,7 @@ mod tests {
 	}
 
 	// used for generating assignments where the validity of the VRF doesn't matter.
-	fn garbage_vrf() -> (VRFOutput, VRFProof) {
+	pub(crate) fn garbage_vrf() -> (VRFOutput, VRFProof) {
 		let key = Sr25519Keyring::Alice.pair();
 		let key: &schnorrkel::Keypair = key.as_ref();
